chore: move getAffiliateAppDataFragmentByChainId to cowswap-utils
