+++ packages/swapper/src/swappers/CowSwapper/getCowSwapTradeQuote/getCowSwapTradeQuote.ts
@@ -8,7 +8,7 @@ import type { AxiosError } from 'axios'
 import { zeroAddress } from 'viem'
 
 import { getDefaultSlippageDecimalPercentageForSwapper } from '../../../constants'
-import { getCowNetwork } from '../../../cowswap-utils'
+import { getAffiliateAppDataFragmentByChainId, getCowNetwork } from '../../../cowswap-utils'
 import type {
   GetEvmTradeQuoteInputBase,
   SwapErrorRight,
@@ -26,7 +26,6 @@ import {
 import { cowService } from '../utils/cowService'
 import {
   assertValidTrade,
-  getAffiliateAppDataFragmentByChainId,
   getFullAppData,
   getNowPlusThirtyMinutesTimestamp,
   getValuesFromQuoteResponse,
