+++ packages/swapper/src/swappers/CowSwapper/getCowSwapTradeRate/getCowSwapTradeRate.ts
@@ -8,7 +8,7 @@ import type { AxiosError } from 'axios'
 import { zeroAddress } from 'viem'
 
 import { getDefaultSlippageDecimalPercentageForSwapper } from '../../../constants'
-import { getCowNetwork } from '../../../cowswap-utils'
+import { getAffiliateAppDataFragmentByChainId, getCowNetwork } from '../../../cowswap-utils'
 import type { GetEvmTradeRateInput, SwapErrorRight, SwapperConfig, TradeRate } from '../../../types'
 import { SwapperName, TradeQuoteError } from '../../../types'
 import { createTradeAmountTooSmallErr, makeSwapErrorRight } from '../../../utils'
@@ -21,7 +21,6 @@ import {
 import { cowService } from '../utils/cowService'
 import {
   assertValidTrade,
-  getAffiliateAppDataFragmentByChainId,
   getFullAppData,
   getNowPlusThirtyMinutesTimestamp,
   getValuesFromQuoteResponse,
